pyproject.toml
[project]
name = "cctop"
version = "0.1.0"
description = "HTOP-like TUI monitor for Claude Code agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "textual>=0.88.0",
    "rich>=13.0.0",
    "watchdog>=5.0.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
]

[project.scripts]
cctop = "cctop.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
